+++ app/src/main/java/com/example/todoapp/ui/theme/extended/ExtendedColors.kt
@@ -1,8 +1,37 @@
-package com.example.todoapp.ui.theme
+package com.example.todoapp.ui.theme.extended
 
 import androidx.compose.runtime.Immutable
 import androidx.compose.runtime.staticCompositionLocalOf
 import androidx.compose.ui.graphics.Color
+import com.example.todoapp.ui.theme.DarkBackElevated
+import com.example.todoapp.ui.theme.DarkBackPrimary
+import com.example.todoapp.ui.theme.DarkBackSecondary
+import com.example.todoapp.ui.theme.DarkColorBlue
+import com.example.todoapp.ui.theme.DarkColorGray
+import com.example.todoapp.ui.theme.DarkColorGrayLight
+import com.example.todoapp.ui.theme.DarkColorGreen
+import com.example.todoapp.ui.theme.DarkColorRed
+import com.example.todoapp.ui.theme.DarkLabelDisable
+import com.example.todoapp.ui.theme.DarkLabelPrimary
+import com.example.todoapp.ui.theme.DarkLabelSecondary
+import com.example.todoapp.ui.theme.DarkLabelTertiary
+import com.example.todoapp.ui.theme.DarkSupportOverlay
+import com.example.todoapp.ui.theme.DarkSupportSeparator
+import com.example.todoapp.ui.theme.LightBackElevated
+import com.example.todoapp.ui.theme.LightBackPrimary
+import com.example.todoapp.ui.theme.LightBackSecondary
+import com.example.todoapp.ui.theme.LightColorBlue
+import com.example.todoapp.ui.theme.LightColorGray
+import com.example.todoapp.ui.theme.LightColorGrayLight
+import com.example.todoapp.ui.theme.LightColorGreen
+import com.example.todoapp.ui.theme.LightColorRed
+import com.example.todoapp.ui.theme.LightLabelDisable
+import com.example.todoapp.ui.theme.LightLabelPrimary
+import com.example.todoapp.ui.theme.LightLabelSecondary
+import com.example.todoapp.ui.theme.LightLabelTertiary
+import com.example.todoapp.ui.theme.LightSupportOverlay
+import com.example.todoapp.ui.theme.LightSupportSeparator
+import com.example.todoapp.ui.theme.White
 
 @Immutable
 data class ExtendedColors(
